Move Locale type from Wire.API.User.Profile to Wire.API.Locale.
