+++ libs/wire-api/src/Wire/API/Provider/External.hs
@@ -27,9 +27,10 @@ import Data.Aeson
 import Data.Id
 import Data.Json.Util ((#))
 import Imports
+import Wire.API.Locale (Locale)
 import Wire.API.Provider.Bot (BotConvView, BotUserView)
 import Wire.API.User.Client.Prekey (LastPrekey, Prekey)
-import Wire.API.User.Profile (Asset, ColourId, Locale, Name)
+import Wire.API.User.Profile (Asset, ColourId, Name)
 import Wire.Arbitrary (Arbitrary, GenericUniform (..))
 
 --------------------------------------------------------------------------------
