+++ libs/wire-api/src/Wire/API/User/Profile.hs
@@ -28,17 +28,6 @@ module Wire.API.User.Profile
     Asset (..),
     AssetSize (..),
 
-    -- * Locale
-    Locale (..),
-    locToText,
-    parseLocale,
-    Language (..),
-    lan2Text,
-    parseLanguage,
-    Country (..),
-    con2Text,
-    parseCountry,
-
     -- * ManagedBy
     ManagedBy (..),
     defaultManagedBy,
@@ -50,19 +39,14 @@ module Wire.API.User.Profile
 where
 
 import Cassandra qualified as C
-import Control.Applicative (optional)
-import Control.Error (hush, note)
+import Control.Error (note)
 import Data.Aeson (FromJSON (..), ToJSON (..))
 import Data.Aeson qualified as A
 import Data.Attoparsec.ByteString.Char8 (takeByteString)
-import Data.Attoparsec.Text
 import Data.ByteString.Conversion
-import Data.ISO3166_CountryCodes
-import Data.LanguageCodes
 import Data.OpenApi qualified as S
 import Data.Range
 import Data.Schema
-import Data.Text qualified as Text
 import Imports
 import Wire.API.Asset (AssetKey (..))
 import Wire.API.User.Orphans ()
@@ -189,87 +173,6 @@ instance C.Cql AssetSize where
   toCql AssetComplete = C.CqlInt 1
 
 --------------------------------------------------------------------------------
--- Locale
-
-data Locale = Locale
-  { lLanguage :: Language,
-    lCountry :: Maybe Country
-  }
-  deriving stock (Eq, Ord, Generic)
-  deriving (Arbitrary) via (GenericUniform Locale)
-  deriving (FromJSON, ToJSON, S.ToSchema) via Schema Locale
-
-instance ToSchema Locale where
-  schema = locToText .= parsedText "Locale" (note err . parseLocale)
-    where
-      err = "Invalid locale. Expected <ISO 639-1>(-<ISO 3166-1-alpha2>)? format"
-
-instance Show Locale where
-  show = Text.unpack . locToText
-
-locToText :: Locale -> Text
-locToText (Locale l c) = lan2Text l <> foldMap (("-" <>) . con2Text) c
-
-parseLocale :: Text -> Maybe Locale
-parseLocale = hush . parseOnly localeParser
-  where
-    localeParser :: Parser Locale
-    localeParser =
-      Locale
-        <$> (languageParser <?> "Language code")
-        <*> (optional (char '-' *> countryParser) <?> "Country code")
-
---------------------------------------------------------------------------------
--- Language
-
-newtype Language = Language {fromLanguage :: ISO639_1}
-  deriving stock (Eq, Ord, Show, Generic)
-  deriving newtype (Arbitrary, S.ToSchema)
-
-instance C.Cql Language where
-  ctype = C.Tagged C.AsciiColumn
-  toCql = C.toCql . lan2Text
-
-  fromCql (C.CqlAscii l) = case parseLanguage l of
-    Just l' -> pure l'
-    Nothing -> Left "Language: ISO 639-1 expected."
-  fromCql _ = Left "Language: ASCII expected"
-
-languageParser :: Parser Language
-languageParser = codeParser "language" $ fmap Language . checkAndConvert isLower
-
-lan2Text :: Language -> Text
-lan2Text = Text.toLower . Text.pack . show . fromLanguage
-
-parseLanguage :: Text -> Maybe Language
-parseLanguage = hush . parseOnly languageParser
-
---------------------------------------------------------------------------------
--- Country
-
-newtype Country = Country {fromCountry :: CountryCode}
-  deriving stock (Eq, Ord, Show, Generic)
-  deriving newtype (Arbitrary, S.ToSchema)
-
-instance C.Cql Country where
-  ctype = C.Tagged C.AsciiColumn
-  toCql = C.toCql . con2Text
-
-  fromCql (C.CqlAscii c) = case parseCountry c of
-    Just c' -> pure c'
-    Nothing -> Left "Country: ISO 3166-1-alpha2 expected."
-  fromCql _ = Left "Country: ASCII expected"
-
-countryParser :: Parser Country
-countryParser = codeParser "country" $ fmap Country . checkAndConvert isUpper
-
-con2Text :: Country -> Text
-con2Text = Text.pack . show . fromCountry
-
-parseCountry :: Text -> Maybe Country
-parseCountry = hush . parseOnly countryParser
-
---------------------------------------------------------------------------------
 -- ManagedBy
 
 -- | Who controls changes to the user profile (where the profile is defined as "all
@@ -357,18 +260,3 @@ instance C.Cql Pict where
 
 noPict :: Pict
 noPict = Pict []
-
---------------------------------------------------------------------------------
--- helpers
-
--- Common language / country functions
-checkAndConvert :: (Read a) => (Char -> Bool) -> String -> Maybe a
-checkAndConvert f t =
-  if all f t
-    then readMaybe (map toUpper t)
-    else fail "Format not supported."
-
-codeParser :: String -> (String -> Maybe a) -> Parser a
-codeParser err conv = do
-  code <- count 2 anyChar
-  maybe (fail err) pure (conv code)
